requirements.txt
cryptography>=41.0.0
pyotp>=2.8.0
click>=8.1.0
typer>=0.9.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-nmap>=0.7.1
requests>=2.31.0
scapy>=2.5.0
